fix(db,app): clear 359 pre-existing clippy warnings, two of them real bugs
`cargo clippy --workspace --all-targets` reported 359 warnings. All are now
gone. The interesting part is not the lints but the two places where a lint
was pointing at code that could not work.

Bugs this fixes:

- `days_to_iso_date` used 146_097 where the Howard Hinnant civil-from-days
  algorithm requires 146_096. The two agree on 19999 of the first 20000
  days, so nothing looked wrong, but every 400-year boundary date came out
  one day late: 2000-02-29 rendered as 2000-03-01. This is the function that
  derives "yesterday" for login streaks, so a streak could extend or reset
  on the wrong day. Two tests now pin it, both confirmed to fail against
  the old divisor.
- `bulk_export` called `store.delete_if_unreferenced(...)` without awaiting
  it. A dropped future never runs, so when `store.reference` failed the
  just-written blob was never deleted and the volume grew by an orphan per
  failed export. Now awaited on the error path, with the cleanup failure
  logged instead of discarded.

Everything else is mechanical, but a few items are worth naming because the
lint was the only thing flagging them:

- Five `pub fn from_str(s: &str) -> Option<Self>` inherent methods in
  `domain` (SanctionLevel, FederationScope, WarningType, WarningAction,
  ThreadMode) shadowed `FromStr` while the ids module already used the trait.
  Converted to real `impl FromStr`, so `"x".parse()` now works on all five and
  the "confused for the standard trait method" lint is gone for good rather
  than suppressed. The domain round-trip tests moved from asserting `None`
  to asserting `Err`.
- `direction.rs:1` and the two `doc list item without indentation` sites had
  doc comments detached from the items they document by a blank line or a
  section banner, so `recompute_all_taste_vectors` was documented by a line
  about a different function. Reattached.
- `&PathBuf` in 13 test helpers became `&Path`, plus 9 `sort_by` closures
  became `sort_by_key`, and the ten-element `bounties` row tuples became a
  named `BountyRow` alias.
- Three genuinely dead test helpers removed (`config_for` x2, a `cookie`
  accessor) along with the imports they alone used.
- 10 DB functions carry `#[allow(clippy::too_many_arguments)]`, matching the
  9 already there. Explicit parameters are the house style; a builder would
  only move the same fields one call deeper.

1708 tests pass. 0 errors, 0 warnings, down from 359.

+++ crates/app/src/routes/imports.rs
@@ -592,7 +592,7 @@ async fn preview_import(
                 .map(|chapter| ChapterIdentity {
                     word_count: None,
                     source_chapter_key: chapter.source_chapter_key.clone(),
-                    ordinal: chapter.ordinal as u32,
+                    ordinal: chapter.ordinal,
                     title: chapter.title.clone(),
                 })
                 .collect(),
@@ -875,7 +875,7 @@ async fn start_import(
                     .iter()
                     .map(|chapter| ChapterIdentity {
                         source_chapter_key: chapter.source_chapter_key.clone(),
-                        ordinal: chapter.ordinal as u32,
+                        ordinal: chapter.ordinal,
                         word_count: None,
                         title: chapter.title.clone(),
                     })
